requirements.txt
google-adk
uvicorn
fastapi
